requirements.txt
# ─── Web Framework ───────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# ─── Google APIs ─────────────────────────────────────────────
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0

# ─── MongoDB (pre-built wheels on Windows) ───────────────────
motor>=3.4.0
pymongo>=4.8.0

# ─── Redis ───────────────────────────────────────────────────
redis>=5.0.0

# ─── HTTP client ─────────────────────────────────────────────
httpx>=0.26.0

# ─── AI / NLP ────────────────────────────────────────────────
google-genai>=0.1.0
dateparser>=1.2.0
pytz>=2024.1

# ─── Data Validation (pydantic v2 with pre-built wheels) ─────
pydantic>=2.5.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# ─── Email Parsing ───────────────────────────────────────────
beautifulsoup4>=4.12.0
html2text>=2024.2.26

# ─── Testing ─────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0

